+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.3.1.1
 
 - Fix add curse pack button not working when you have a link copied
+- Add Open CurseForge button on Instances tab for Curse packs